% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Confidence Bootstrap Interval}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations to be done.}

\item{x}{data set to sample.}

\item{fun}{function to estimate.}

\item{alpha}{error.}

\item{cx}{point size for graph.}

\item{...}{any other relevant parameters.}
}
\value{
returns a list of data containing the point estimate and the interval.
}
\description{
Calculates a confidence interval based on resampling with an alpha (error) value, number of iteratorations (time to resample).
Also, estimates based on the function passed in with the sample x.
Finally, plotting the estimation interval.
}
\examples{
\dontrun{myboot2(x=FALL234753ghaz0010::fire$DISTANCE)}
}
