% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{Central Limit Theorem Binomial Function}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{the size.}

\item{iter}{the number of iterations.}

\item{p}{the probability.}

\item{...}{any remaining parameters.}
}
\value{
a vector containing the x-values of the plot (sample mean) vs. the y-values of the plot (density).
}
\description{
{Application of the CLT with discrete and continuous
binomial approximation.}
}
\examples{
\dontrun{mycltb(n=10, iter=10000, p=0.5)}
}
