% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{myfun}
\alias{myfun}
\title{My first function}
\usage{
myfun(x)
}
\arguments{
\item{x}{A quantative vector containing values.}
}
\value{
A vector of squared components.
}
\description{
My first function
}
\examples{
myfun(1:10)
}
