% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-estimator.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum Likelihood Estimator}
\usage{
mymaxlik(lfun = logbin, x, param, ...)
}
\arguments{
\item{lfun}{the function to be used.}

\item{x}{the list of values we want to find an approx for.}

\item{param}{similar to bounds.}

\item{...}{any other relevant parameters (title, etc.).}
}
\value{
returns a plot of the MLE estimator and a list of data.
}
\description{
{Used for repeated sampling from the same distribution, and
calculates the maximum likelihood estimator.}
}
\examples{
\dontrun{mymaxlik(x=c(3,3,4,3,4,5,5,4),param=seq(0,1,length=1000),
\dontrun lfun=logbin,xlab=expression(pi),main="Binomial",cex.main=2)}
}
