% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negativeBinom.R
\name{mynbin}
\alias{mynbin}
\title{Negative Binomial}
\usage{
mynbin(y, r, p)
}
\arguments{
\item{y}{the number of trials until the 'rth' success is observed.}

\item{r}{the size / number of total trials.}

\item{p}{the probability.}
}
\value{
returns the negative binomial distribution data.
}
\description{
{The following function will calculate the negative binomial distribution by pure definition.
That is, it will utilize the choose function. For example (P<=10).}
}
\examples{
\dontrun{mynbin(y=10, r=3, p=0.4)}
}
