% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal--continous--distribution.R
\name{myncurve}
\alias{myncurve}
\title{Normal Distribution Plot}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the expected (mean) value.}

\item{sigma}{the expected standard deviation value.}

\item{a}{the "y" value to check, such that P(Y <= a).}
}
\value{
returns the normal distribution plotted, shaded for the area under the curve, given a, and the P(Y<=a).
}
\description{
Normal Distribution Plot
}
