% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull--reparameterization--r.R
\name{reparwbl}
\alias{reparwbl}
\title{Weibull reparameterization from the book to R.}
\usage{
reparwbl(a, b)
}
\arguments{
\item{a}{alpha value given from the book.}

\item{b}{beta value given from the book.}
}
\value{
returns a list of the correct R parameterization.
}
\description{
This function reparameterizes the weibull distribution into R.
}
