% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{Central Limit Theorem Binomial Function

The function works for discrete and continous distributions.}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{the size.}

\item{iter}{the number of iterations.}

\item{p}{the probability.}

\item{...}{any remaning parameters.}
}
\value{
a vector containg the x-values of the plot (sample mean) vs. the y-values of the plot (density).
}
\description{
Central Limit Theorem Binomial Function

The function works for discrete and continous distributions.
}
\examples{
mycltb(n=10, iter=10000, p=0.5)
}
