% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{myfun}
\alias{myfun}
\title{x-squared}
\usage{
myfun(x)
}
\arguments{
\item{x}{A quantitative vector containing values.}
}
\value{
A vector of squared components.
}
\description{
Squares the x-term of the vector (or single value).
}
\examples{
\dontrun{myfun(1:10)}
}
