% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negativeBinom.R
\name{mynbin}
\alias{mynbin}
\title{Negative Binomial}
\usage{
mynbin(y, r, p)
}
\arguments{
\item{y}{the number of trials until the 'rth' sucess is observed.}

\item{r}{the size / number of total trials.}

\item{p}{the probability.}
}
\value{
returns the negative binomial distribution data.
}
\description{
Negative Binomial
}
\examples{
#P(Y=10)Y~NegBin(p=0.4,r=3)
}
