% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynticketsOverbook.R
\name{ntickets}
\alias{ntickets}
\title{Overbooking Ticket Function (CLT)

Based on the Central limit theorm, we can use a continous normal approximation for a discrete variable.
Thus, we solve for -n- number of seats sold, with the possibility of an overbook in both the discrete case and the continuous case.
We used functions such as which.min() and stats::uniroot() to calculate these values.
From there, we calculated nc and nd as-well as the other given values and put them into a list.
Finally, we plotted both the discrete and continous approximations.}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{the number of seats on the flight.}

\item{gamma}{the probability a plane will be truly overbooked.}

\item{p}{the probability of showing up.}
}
\value{
returns multiple plots, a list of the discrete and continuous approximations w/ other parameters.
}
\description{
Overbooking Ticket Function (CLT)

Based on the Central limit theorm, we can use a continous normal approximation for a discrete variable.
Thus, we solve for -n- number of seats sold, with the possibility of an overbook in both the discrete case and the continuous case.
We used functions such as which.min() and stats::uniroot() to calculate these values.
From there, we calculated nc and nd as-well as the other given values and put them into a list.
Finally, we plotted both the discrete and continous approximations.
}
