% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pieceWiseIndicator.R
\name{pieceWiseFunction}
\alias{pieceWiseFunction}
\title{Piecewise function checker.}
\usage{
pieceWiseFunction(x, coef, xk)
}
\arguments{
\item{x}{the x-vector containg points to be checked or just one x-value.}

\item{coef}{the coefficents of the model (b0, b1, b2..)}

\item{xk}{the point at which both lines meet up.}
}
\value{
returns the a vector of y-points based on the equation or just one y-value.
}
\description{
This function will check if a function follows a piece-wise model.
}
\examples{
\dontrun{pieceWiseFunction(x=1, coef=2, xk=3)}
}
